Fix legion dim in gather

+++ lib/local-execution/include/legion_tensor_shape.h
@@ -28,8 +28,8 @@ public:
   DataType data_type;
 };
 
-ff_dim_t to_ff(legion_dim_t, int num_dims);
-legion_dim_t to_legion(ff_dim_t, int num_dims);
+ff_dim_t to_ff(legion_dim_t, size_t num_dims);
+legion_dim_t to_legion(ff_dim_t, size_t num_dims);
 
 ff_dim_t to_ff(legion_dim_t, TensorShape const &);
 legion_dim_t to_legion(ff_dim_t, TensorShape const &);
